+++ bika/health/content/analysisrequest.py
@@ -82,7 +82,9 @@ class AnalysisRequestSchemaExtender(object):
                     # UID is required in colModel
                     {'columnName': 'UID', 'hidden': True},
                 ],
-                showOn=True,
+                showOn=False,
+                delay=1000,
+                minLength=2,
                 add_button={
                     'visible': True,
                     'url': 'patients/portal_factory/Patient/new/edit',