+++ CHANGES.rst
@@ -18,6 +18,7 @@ Changelog
 
 - #72 Slow searches in patients list
 - #71 Slow Reference Widgets in 'Add Case' View
+- #73 Slow Reference Widgets in 'Add AR' View
 - #69 Add proper license to repository
 
 **Security**